separate NIST and NIST-TA3 restrictions
