+++ java/src/test/java/com/ncc/aif/NistExamplesAndValidationTest.java
@@ -69,6 +69,9 @@ public class NistExamplesAndValidationTest {
        ((Logger) org.slf4j.LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME)).setLevel(Level.INFO);
        Set<String> ont = Set.of("com/ncc/aif/ontologies/LDCOntology");
        utils = new NistTestUtils(LDC_NS, ValidateAIF.create(ont, ValidateAIF.Restriction.NIST), DUMP_ALWAYS, DUMP_TO_FILE);
+       
+       //utils = new NistTestUtils(LDC_NS, ValidateAIF.createForDWD(ValidateAIF.Restriction.NIST), DUMP_ALWAYS,
+       //                                 DUMP_TO_FILE);
    }
  
    private Model model;
@@ -925,202 +928,202 @@ public class NistExamplesAndValidationTest {
            }
        }
  
-       @Nested
-       class ValidExamples {
-               Resource validProtoType;
-               Resource validSameAsCluster;
+    //    @Nested
+    //    class ValidExamples {
+    //            Resource validProtoType;
+    //            Resource validSameAsCluster;
  
-               /**
-                * Create ClaimComponent objects
-                */      
-               @Nested
-               class ClaimComponentTest {
-                   String claimComponentURI = "https://www.wikidata.org/wiki/Q8440";
+    //            /**
+    //             * Create ClaimComponent objects
+    //             */      
+    //            @Nested
+    //            class ClaimComponentTest {
+    //                String claimComponentURI = "https://www.wikidata.org/wiki/Q8440";
  
-                   ClaimComponent validComponent = new ClaimComponent()
-                                   .setName("Hugo Chávez")
-                                   .setIdentity("Q8440")
-                                   .addType("Q82955"); // Politician
+    //                ClaimComponent validComponent = new ClaimComponent()
+    //                                .setName("Hugo Chávez")
+    //                                .setIdentity("Q8440")
+    //                                .addType("Q82955"); // Politician
  
  
-                   @Test
-                   void validMininmal() {
-                           validComponent.addToModel(model, utils.getUri("a_valid_minimal_claimcomponent"), system);
-                           //utils.testValid("Create minimal valid ClaimComponent");
-                           utils.testValid("Create minimal valid");                           
-                   }
+    //                @Test
+    //                void validMininmal() {
+    //                        validComponent.addToModel(model, utils.getUri("a_valid_minimal_claimcomponent"), system);
+    //                        //utils.testValid("Create minimal valid ClaimComponent");
+    //                        utils.testValid("Create minimal valid");                           
+    //                }
  
-                   @Test
-                   void validFull() {
+    //                @Test
+    //                void validFull() {
                           
-                           // Resource validProtoType1 = makeEntity(model, utils.getUri("someTestURI1"), system);                                       
-                           // Resource validSameAsCluster1 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID1", InterchangeOntology.SameAsCluster, system)
-                           //         .addProperty(InterchangeOntology.prototype, validProtoType1);
+    //                        // Resource validProtoType1 = makeEntity(model, utils.getUri("someTestURI1"), system);                                       
+    //                        // Resource validSameAsCluster1 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID1", InterchangeOntology.SameAsCluster, system)
+    //                        //         .addProperty(InterchangeOntology.prototype, validProtoType1);
                                   
-                           validComponent.setProvenance("Hugo Chavez")
-                                   //.setKE(validSameAsCluster1)
-                                   .addToModel(model, utils.getUri("a_valid_full_claimcomponent"), system);
-                           utils.testValid("Create full valid ClaimComponent");
-                   }
+    //                        validComponent.setProvenance("Hugo Chavez")
+    //                                //.setKE(validSameAsCluster1)
+    //                                .addToModel(model, utils.getUri("a_valid_full_claimcomponent"), system);
+    //                        utils.testValid("Create full valid ClaimComponent");
+    //                }
  
-                   @Test
-                   void invalidMissingType() {
-                           validComponent.setTypes(Collections.emptySet()) // remove types
-                                   .addToModel(model, utils.getUri("an_invalid_missing_type"), system);
-                           utils.expect(null, SH.MinCountConstraintComponent, null);
-                           utils.testInvalid(
-                                           "ClaimComponent.invalid (missing type): ClaimComponent must have a type");
-                   }
+    //                @Test
+    //                void invalidMissingType() {
+    //                        validComponent.setTypes(Collections.emptySet()) // remove types
+    //                                .addToModel(model, utils.getUri("an_invalid_missing_type"), system);
+    //                        utils.expect(null, SH.MinCountConstraintComponent, null);
+    //                        utils.testInvalid(
+    //                                        "ClaimComponent.invalid (missing type): ClaimComponent must have a type");
+    //                }
  
-                   @Test
-                   void invalidTooManyTypes(){
+    //                @Test
+    //                void invalidTooManyTypes(){
  
-                       //Test max type count of 5   
-                       validComponent.addType("Q829551")
-                                     .addType("Q829552")
-                                     .addType("Q829553")
-                                     .addType("Q829554")
-                                     .addType("Q829555")
-                                     .addType("Q829556")                                      
-                                     .addType("Q829557")
-                                     .addToModel(model, utils.getUri("an_invalid_too_many_types"), system);
+    //                    //Test max type count of 5   
+    //                    validComponent.addType("Q829551")
+    //                                  .addType("Q829552")
+    //                                  .addType("Q829553")
+    //                                  .addType("Q829554")
+    //                                  .addType("Q829555")
+    //                                  .addType("Q829556")                                      
+    //                                  .addType("Q829557")
+    //                                  .addToModel(model, utils.getUri("an_invalid_claimcomponent_too_many_types"), system);
 
-                       utils.expect(null, SH.BlankNode, null);
-                       utils.testInvalid(
-                               "ClaimComponent.invalid (Too many type): ClaimComponent must max 5 types");
+    //                    utils.expect(null, SH.MaxCountConstraintComponent, null);
+    //                    utils.testInvalid(
+    //                            "ClaimComponent.invalid (Too many type): ClaimComponent must max 5 types");
         
-                   }
-               }
+    //                }
+    //            }
  
-               /**
-                * Create ClaimComponent objects
-                */
-               @Nested
-               class ClaimTest {
-                       Resource validXComponent;
-                       Resource validComponentKE;
-                       Resource validClaimerComponent;
-                       Resource validClaimLocationComponent;
-                       Resource validProtoType1;
-                       Resource validProtoType2;
-                       Resource validProtoType3;
-                       Resource validSameAsCluster1;
-                       Resource validSameAsCluster2;
-                       Resource validSameAsCluster3;
-                       Claim validClaim;
+    //            /**
+    //             * Create ClaimComponent objects
+    //             */
+    //            @Nested
+    //            class ClaimTest {
+    //                    Resource validXComponent;
+    //                    Resource validComponentKE;
+    //                    Resource validClaimerComponent;
+    //                    Resource validClaimLocationComponent;
+    //                    Resource validProtoType1;
+    //                    Resource validProtoType2;
+    //                    Resource validProtoType3;
+    //                    Resource validSameAsCluster1;
+    //                    Resource validSameAsCluster2;
+    //                    Resource validSameAsCluster3;
+    //                    Claim validClaim;
  
-                       @BeforeEach
-                       void setup() {
+    //                    @BeforeEach
+    //                    void setup() {
  
-                           validProtoType1 = makeEntity(model, utils.getUri("someTestURI1"), system);
-                           validProtoType2 = makeEntity(model, utils.getUri("someTestURI2"), system);
-                           validProtoType3 = makeEntity(model, utils.getUri("someTestURI3"), system);                           
+    //                        validProtoType1 = makeEntity(model, utils.getUri("someTestURI1"), system);
+    //                        validProtoType2 = makeEntity(model, utils.getUri("someTestURI2"), system);
+    //                        validProtoType3 = makeEntity(model, utils.getUri("someTestURI3"), system);                           
  
-                           validSameAsCluster1 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID1", InterchangeOntology.SameAsCluster, system)
-                           .addProperty(InterchangeOntology.prototype, validProtoType1);
-                           validSameAsCluster2 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID2", InterchangeOntology.SameAsCluster, system)
-                                   .addProperty(InterchangeOntology.prototype, validProtoType2);
-                           validSameAsCluster3 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID3", InterchangeOntology.SameAsCluster, system)
-                                   .addProperty(InterchangeOntology.prototype, validProtoType3);
+    //                        validSameAsCluster1 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID1", InterchangeOntology.SameAsCluster, system)
+    //                        .addProperty(InterchangeOntology.prototype, validProtoType1);
+    //                        validSameAsCluster2 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID2", InterchangeOntology.SameAsCluster, system)
+    //                                .addProperty(InterchangeOntology.prototype, validProtoType2);
+    //                        validSameAsCluster3 = AIFUtils.makeAIFResource(model, "http://www.caci.com/cluster/SameAsCluster/ClusterID3", InterchangeOntology.SameAsCluster, system)
+    //                                .addProperty(InterchangeOntology.prototype, validProtoType3);
  
-                           //validComponentKE = utils.makeValidAIFEntity(LDCOntology.PER, utils.getUri("pointer_to_some_ke_arg"));
-                           validComponentKE = validSameAsCluster1;                            
+    //                        //validComponentKE = utils.makeValidAIFEntity(LDCOntology.PER, utils.getUri("pointer_to_some_ke_arg"));
+    //                        validComponentKE = validSameAsCluster1;                            
  
-                               validXComponent = new ClaimComponent()
-                                               .setName("Some Agency")
-                                               .setIdentity("Q37230")
-                                               .addType("Q47913") // Politician
-                                               //.setKE(validComponentKE)
-                                               .addToModel(model, "https://www.wikidata.org/wiki/Q37230", system);
+    //                            validXComponent = new ClaimComponent()
+    //                                            .setName("Some Agency")
+    //                                            .setIdentity("Q37230")
+    //                                            .addType("Q47913") // Politician
+    //                                            //.setKE(validComponentKE)
+    //                                            .addToModel(model, "https://www.wikidata.org/wiki/Q37230", system);
  
-                               validClaimerComponent = new ClaimComponent()
-                                               .setName("Some News Outlet")
-                                               .setIdentity("Q48340")
-                                               .addType("Q7892363") // Politician
-                                               //.setKE(validComponentKE)
-                                               .addToModel(model, "https://www.wikidata.org/wiki/Q48340", system);
+    //                            validClaimerComponent = new ClaimComponent()
+    //                                            .setName("Some News Outlet")
+    //                                            .setIdentity("Q48340")
+    //                                            .addType("Q7892363") // Politician
+    //                                            //.setKE(validComponentKE)
+    //                                            .addToModel(model, "https://www.wikidata.org/wiki/Q48340", system);
       
-                               validClaimLocationComponent = new ClaimComponent()
-                                               .setName("Some Country")
-                                               .setIdentity("Q717")
-                                               .addType("Q3624078") // Politician
-                                               //.setKE(validComponentKE)
-                                               .addToModel(model, "https://www.wikidata.org/wiki/Q717", system);                
+    //                            validClaimLocationComponent = new ClaimComponent()
+    //                                            .setName("Some Country")
+    //                                            .setIdentity("Q717")
+    //                                            .addType("Q3624078") // Politician
+    //                                            //.setKE(validComponentKE)
+    //                                            .addToModel(model, "https://www.wikidata.org/wiki/Q717", system);                
                           
  
-                               validClaim = new Claim()
-                                               .setSourceDocument("Some source")
-                                               .setTopic("Some Main Topic: Death of Hugo Chavez")
-                                               .setSubtopic("Some Sub TubTopic: Who killed Hugo Chavez")
-                                               .setClaimTemplate("X killed Hugo Chavez")
-                                               .addXVariable(validXComponent)
-                                               .setNaturalLanguageDescription("Claimer Y claims X killed Hugo Chavez")
-                                               .addClaimSementics(validSameAsCluster1)
-                                               .setClaimer(validClaimerComponent)
-                                               .addAssociatedKE(validSameAsCluster2)
-                                               .addAssociatedKE(validSameAsCluster3);
-                       }
+    //                            validClaim = new Claim()
+    //                                            .setSourceDocument("Some source")
+    //                                            .setTopic("Some Main Topic: Death of Hugo Chavez")
+    //                                            .setSubtopic("Some Sub TubTopic: Who killed Hugo Chavez")
+    //                                            .setClaimTemplate("X killed Hugo Chavez")
+    //                                            .addXVariable(validXComponent)
+    //                                            .setNaturalLanguageDescription("Claimer Y claims X killed Hugo Chavez")
+    //                                            .addClaimSementics(validSameAsCluster1)
+    //                                            .setClaimer(validClaimerComponent)
+    //                                            .addAssociatedKE(validSameAsCluster2)
+    //                                            .addAssociatedKE(validSameAsCluster3);
+    //                    }
  
-                       @Test
-                       void validMinimal() {
-                               validClaim.addToModel(model, utils.getUri("a_minimal_claimframe"), system);
-                               utils.testValid("Create minimal valid claim frame");
-                       }
+    //                    @Test
+    //                    void validMinimal() {
+    //                            validClaim.addToModel(model, utils.getUri("a_minimal_claimframe"), system);
+    //                            utils.testValid("Create minimal valid claim frame");
+    //                    }
  
-                       @Test
-                       void validFull() {
-                               // Resource someOtherClaimFrame1 = model.createResource("https://www.caci.com/claim/someOtherClaimID1");
-                               // Resource someOtherClaimFrame2 = model.createResource("https://www.caci.com/claim/someOtherClaimID2");
-                               // Resource someOtherClaimFrame3 = model.createResource("https://www.caci.com/claim/someOtherClaimID3");
+    //                    @Test
+    //                    void validFull() {
+    //                            // Resource someOtherClaimFrame1 = model.createResource("https://www.caci.com/claim/someOtherClaimID1");
+    //                            // Resource someOtherClaimFrame2 = model.createResource("https://www.caci.com/claim/someOtherClaimID2");
+    //                            // Resource someOtherClaimFrame3 = model.createResource("https://www.caci.com/claim/someOtherClaimID3");
  
-                               String someOtherClaimFrame1 = "someOtherClaimID1";
-                               String someOtherClaimFrame2 = "someOtherClaimID2";
-                               String someOtherClaimFrame3 = "someOtherClaimID3";
+    //                            String someOtherClaimFrame1 = "someOtherClaimID1";
+    //                            String someOtherClaimFrame2 = "someOtherClaimID2";
+    //                            String someOtherClaimFrame3 = "someOtherClaimID3";
                               
  
-                               validClaim
-                                   .setImportance(1d)
-                                   .setClaimId("claimId")
-                                   .setQueryId("queryId")
-                                   .setClaimLocation(validClaimLocationComponent)
-                                   .setClaimMedium(validClaimLocationComponent)
-                                   .addClaimerAfilliation(validClaimerComponent)
-                                   .addIdenticalClaim(someOtherClaimFrame1)
-                                   .addRelatedClaim(someOtherClaimFrame2)
-                                   .addSupportingClaim(someOtherClaimFrame1)
-                                   .addSupportingClaim(someOtherClaimFrame2)
-                                   .addRefutingClaim(someOtherClaimFrame1)
-                                   .addRefutingClaim(someOtherClaimFrame2)
-                                   .addRefutingClaim(someOtherClaimFrame3);
+    //                            validClaim
+    //                                .setImportance(1d)
+    //                                .setClaimId("claimId")
+    //                                .setQueryId("queryId")
+    //                                .setClaimLocation(validClaimLocationComponent)
+    //                                .setClaimMedium(validClaimLocationComponent)
+    //                                .addClaimerAfilliation(validClaimerComponent)
+    //                                .addIdenticalClaim(someOtherClaimFrame1)
+    //                                .addRelatedClaim(someOtherClaimFrame2)
+    //                                .addSupportingClaim(someOtherClaimFrame1)
+    //                                .addSupportingClaim(someOtherClaimFrame2)
+    //                                .addRefutingClaim(someOtherClaimFrame1)
+    //                                .addRefutingClaim(someOtherClaimFrame2)
+    //                                .addRefutingClaim(someOtherClaimFrame3);
                                   
-                               validClaim.setClaimDateTime(AIFUtils.makeLDCTimeRange(model,
-                                   "2013-01-xx", "2013-12-xx", "2014-01-xx", "2014-12-xx", system));
+    //                            validClaim.setClaimDateTime(AIFUtils.makeLDCTimeRange(model,
+    //                                "2013-01-xx", "2013-12-xx", "2014-01-xx", "2014-12-xx", system));
  
-                               validClaim.addToModel(model, utils.getUri("a_full_claimframe"), system);
+    //                            validClaim.addToModel(model, utils.getUri("a_full_claimframe"), system);
  
-                               utils.testValid("Create full valid claim frame");
-                       }
+    //                            utils.testValid("Create full valid claim frame");
+    //                    }
  
 
-                       @Test
-                       void invalidMissingClaimID(){
-                           validClaim
-                           .setImportance(1d)
-                           .setClaimId("claimId")      
-                           .setClaimTemplate(null);
+    //                    @Test
+    //                    void invalidMissingClaimID(){
+    //                        validClaim
+    //                        .setImportance(1d)
+    //                        .setClaimId("claimId")      
+    //                        .setClaimTemplate(null);
                           
-                       }
+    //                    }
 
-                       // @Test (this is now valid outside of restricted set)
-                       // void invalidMissingXVariable() {
-                       //         validClaim.setXVariable(Collections.emptySet()).addToModel(model, utils.getUri("claim"),
-                       //                         system);
-                       //         utils.expect(null, SH.MinCountConstraintComponent, null);
-                       //         utils.testInvalid("ClaimTest.invalid (missing x variable): Claim must have X Variable");
-                       // }
-               }
+    //                    // @Test (this is now valid outside of restricted set)
+    //                    // void invalidMissingXVariable() {
+    //                    //         validClaim.setXVariable(Collections.emptySet()).addToModel(model, utils.getUri("claim"),
+    //                    //                         system);
+    //                    //         utils.expect(null, SH.MinCountConstraintComponent, null);
+    //                    //         utils.testInvalid("ClaimTest.invalid (missing x variable): Claim must have X Variable");
+    //                    // }
+    //            }
   
-       }
+    //    }
    }
  
    private void addNamespacesToModel(Model model) {
